Add personality: Change name to Zeus
