--- src/main/java/duke/command/Commands.java
@@ -1,8 +1,0 @@
-package duke.command;
-
-/**
- * Enums for duke
- */
-public enum Commands {
-    TODO, DEADLINE, EVENT
-}
