+++ src/main/java/zeus/command/Commands.java
@@ -1,0 +1,8 @@
+package zeus.command;
+
+/**
+ * Enums for duke
+ */
+public enum Commands {
+    TODO, DEADLINE, EVENT
+}
