+++ src/main/java/zeus/gui/Launcher.java
@@ -1,0 +1,12 @@
+package zeus.gui;
+
+import javafx.application.Application;
+
+/**
+ * A launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
